feat(CB-154): account 정보 및 setting에 따라 전문가 모드 전환 및 등록
- account 정보 및 setting에 따라 전문가 모드 전환 및 등록
- 세팅들 분리
