+++ lib/routes/mainRoute.dart
@@ -42,7 +42,7 @@ class MainRoute extends ConsumerWidget {
       },
       child: Navigator(
         key: _navigatorKey,
-        initialRoute: false ? HOME_PAGE_PATH : '/testpage',
+        initialRoute: true ? HOME_PAGE_PATH : '/testpage',
         //ref.read(tokenProvider.notifier).isEmpty() ? '/auth' : '/home',
         onGenerateRoute: mainRoute,
       ),
